requirements.txt
langchain
pdf2image

streamlit
google-generativeai
python-dotenv
langchain
langchain_google_genai